Update Tag Controller, add Tag Policy

+++ app/Http/Controllers/TagController.php
@@ -7,30 +7,41 @@ namespace App\Http\Controllers;
 use App\Http\Requests\TagRequest;
 use App\Http\Resources\TagResource;
 use App\Models\Tag;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\JsonResponse;
 
 class TagController extends Controller
 {
+    use AuthorizesRequests;
+
     public function store(TagRequest $request): TagResource
     {
+        $this->authorize("store", Tag::class);
         $tag = Tag::query()->create($request->validated());
 
-        return new TagResource($tag);
+        return new TagResource($tag)->additional([
+            "message" => "Tag created successfully",
+        ]);
     }
 
     public function update(TagRequest $request, Tag $tag): TagResource
     {
-        $tag->update([
-            "name" => $request->name,
-        ]);
+        $this->authorize("update", $tag);
 
-        return new TagResource($tag);
+        $tag->update($request->validated());
+
+        return new TagResource($tag)->additional([
+            "message" => "Tag updated successfully",
+        ]);
     }
 
     public function destroy(Tag $tag): JsonResponse
     {
+        $this->authorize("delete", $tag);
         $tag->delete();
 
-        return response()->json(null, 204);
+        return response()->json([
+            "message" => "Tag deleted successfully",
+        ]);
     }
 }
